+++ jhove-core/src/main/java/edu/harvard/hul/ois/jhove/handler/TextHandler.java
@@ -1075,10 +1075,14 @@ public class TextHandler
         _writer.println (margn2 + "TimeBase: 1000");
         _writer.println (margn2 + "VideoField: FIELD_1");
         _writer.println (margn2 + "CountingMode: NTSC_NON_DROP_FRAME");
-        _writer.println (margn2 + "Hours: " + Integer.toString (start.getHours ()));
-        _writer.println (margn2 + "Minutes: " + Integer.toString (start.getMinutes ()));
-        _writer.println (margn2 + "Seconds: " + Integer.toString (start.getSeconds ()));
-        _writer.println (margn2 + "Frames: " + Integer.toString (start.getFrames ()));
+        _writer.println (margn2 + "Hours: " +
+                Long.toString (start.getHours ()));
+        _writer.println (margn2 + "Minutes: " +
+                Long.toString (start.getMinutes ()));
+        _writer.println (margn2 + "Seconds: " +
+                Long.toString (start.getSeconds ()));
+        _writer.println (margn2 + "Frames: " +
+                Long.toString (start.getFrames ()));
         _writer.println (margn2 + "Samples: ");
 	double sr = start.getSampleRate ();
 	if (sr == 1.0) {
@@ -1086,10 +1090,10 @@ public class TextHandler
 	}
 	_writer.println (margn3 + "SampleRate: S" +
 			 Integer.toString ((int) sr));
-	_writer.println (margn3 + "NumberOfSamples: " +
-			 Integer.toString (start.getSamples ()));
+        _writer.println (margn3 + "NumberOfSamples: " +
+                Long.toString (start.getSamples ()));
         _writer.println (margn2 + "FilmFraming: NOT_APPLICABLE");
-	_writer.println (margn3 + "Type: ntscFilmFramingType");
+        _writer.println (margn3 + "Type: ntscFilmFramingType");
 
         if (duration != null) {
            _writer.println (margn1 + "Duration:");
@@ -1098,13 +1102,13 @@ public class TextHandler
            _writer.println (margn2 + "VideoField: FIELD_1");
            _writer.println (margn2 + "CountingMode: NTSC_NON_DROP_FRAME");
            _writer.println (margn2 + "Hours: " +
-			    Integer.toString (duration.getHours ()));
+                    Long.toString (duration.getHours ()));
            _writer.println (margn2 + "Minutes: " +
-			    Integer.toString (duration.getMinutes ()));
+                    Long.toString (duration.getMinutes ()));
            _writer.println (margn2 + "Seconds: " +
-			    Integer.toString (duration.getSeconds ()));
+                    Long.toString (duration.getSeconds ()));
            _writer.println (margn2 + "Frames: " +
-			    Integer.toString (duration.getFrames ()));
+                    Long.toString (duration.getFrames ()));
            _writer.println (margn2 + "Samples: ");
 	   sr = duration.getSampleRate ();
 	   if (sr == 1.0) {
@@ -1112,9 +1116,9 @@ public class TextHandler
 	   }
 	   _writer.println (margn3 + "SampleRate: S" +
 			    Integer.toString ((int) sr));
-	   _writer.println (margn3 + "NumberOfSamples: " +
-			    Integer.toString (duration.getSamples ()));
-           _writer.println (margn2 + "FilmFraming: NOT_APPLICABLE");
+	   _writer.println (margn3 + "NumberOfSamples: " + 
+			    Long.toString (duration.getSamples ()));
+     _writer.println (margn2 + "FilmFraming: NOT_APPLICABLE");
 	   _writer.println (margn3 + "Type: ntscFilmFramingType");
         }
     }
